Refactor: move method 'getGameData' from GameSelectorModel to HeadlessGameServer
