+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/panels/main/game/selector/GameSelectorModel.java
@@ -99,15 +99,6 @@ public class GameSelectorModel extends Observable implements GameSelector {
     load(newData, file.getName());
   }
 
-  public GameData getGameData(final InputStream input) {
-    try {
-      return GameDataManager.loadGame(input);
-    } catch (final IOException e) {
-      log.log(Level.SEVERE, "Failed to load game", e);
-      return null;
-    }
-  }
-
   public void setCanSelect(final boolean canSelect) {
     this.canSelect = canSelect;
     notifyObs();
